+++ libraries/session/src/main/java/androidx/media3/session/MediaSessionImpl.java
@@ -39,6 +39,7 @@ import android.app.PendingIntent;
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.DeadObjectException;
@@ -53,9 +54,11 @@ import android.support.v4.media.session.MediaSessionCompat;
 import android.view.KeyEvent;
 import android.view.ViewConfiguration;
 import androidx.annotation.CheckResult;
+import androidx.annotation.DoNotInline;
 import androidx.annotation.FloatRange;
 import androidx.annotation.GuardedBy;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.media.MediaBrowserServiceCompat;
 import androidx.media3.common.AudioAttributes;
 import androidx.media3.common.DeviceInfo;
@@ -1123,14 +1126,16 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     }
     // Double tap detection.
     int keyCode = keyEvent.getKeyCode();
+    boolean isTvApp = Util.SDK_INT >= 21 && Api21.isTvApp(context);
     boolean doubleTapCompleted = false;
     switch (keyCode) {
       case KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE:
       case KeyEvent.KEYCODE_HEADSETHOOK:
-        if (callerInfo.getControllerVersion() != ControllerInfo.LEGACY_CONTROLLER_VERSION
+        if (isTvApp
+            || callerInfo.getControllerVersion() != ControllerInfo.LEGACY_CONTROLLER_VERSION
             || keyEvent.getRepeatCount() != 0) {
-          // Double tap detection is only for media button events from external sources
-          // (for instance Bluetooth) and excluding long press (repeatCount > 0).
+          // Double tap detection is only for mobile apps that receive a media button event from
+          // external sources (for instance Bluetooth) and excluding long press (repeatCount > 0).
           mediaPlayPauseKeyHandler.flush();
         } else if (mediaPlayPauseKeyHandler.hasPendingPlayPauseTask()) {
           // A double tap arrived. Clear the pending playPause task.
@@ -1815,6 +1820,14 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
       if (!hasMessages(MSG_PLAYER_INFO_CHANGED)) {
         sendEmptyMessage(MSG_PLAYER_INFO_CHANGED);
       }
+    }
+  }
+
+  @RequiresApi(21)
+  private static final class Api21 {
+    @DoNotInline
+    public static boolean isTvApp(Context context) {
+      return context.getPackageManager().hasSystemFeature(PackageManager.FEATURE_LEANBACK);
     }
   }
 }